+++ code/isolation_finder.h
@@ -36,7 +36,7 @@
 struct IsolationRecord {
   bool foundHigherGround;
   LatLng closestHigherGround;
-  float distance;  // distance to peak in meters
+  double distance;  // distance to peak in meters
 
   IsolationRecord()
       : foundHigherGround(false),
@@ -81,7 +81,7 @@ private:
   // Check the neighboring tile with the given lat/lng, where seedCoords give the closest point
   // in the neighboring tile to the peak, and elev is the height of the peak.
   // peakLocation has the same meaning as in findIsolation
-  IsolationRecord checkNeighboringTile(float lat, float lng, const LatLng *peakLocation,
+  IsolationRecord checkNeighboringTile(double lat, double lng, const LatLng *peakLocation,
                                        Offsets seedCoords, Elevation elev) const;
 };
